Formatted code and moved select_all to mod.rs
